requirements.txt
tifffile
numpy
matplotlib
PyTorch
opencv-python
albumentations
sklearn